requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
python-multipart==0.0.9
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
cloudinary==1.40.0
psycopg2-binary==2.9.9
mangum==0.17.0